Rename queue_producer to producer
